requirements.txt
aiohttp==3.11.18
apprise==1.9.3
beautifulsoup4==4.13.4
celery==5.5.2
croniter==6.0.0
Django==5.2
django-allauth[socialaccount]==65.8.1
django-celery-beat==2.8.1
django-celery-results==2.6.0
django_debug_toolbar==5.2.0
django-decorator-include==3.3
django-health-check==3.18.3
django-model-utils==5.0.0
django-redis==5.4.0
django-select2==8.4.0
django-simple-history==3.8.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
icalendar==6.3.1
Pillow==11.2.1
psycopg[binary]==3.2.9
python-decouple==3.8
redis[hiredis]==6.1.0
requests==2.32.3
requests-ratelimiter==0.7.0
unidecode==1.4.0
